PR #1013: Fixing PVS-Studio warnings. (Contributed by PaulEremeeff) Some files have been omitted and will be submitted with modified corrections
[CL 2505544 by Dan Oconnor in Main branch]

+++ Engine/Source/Runtime/CoreUObject/Private/Serialization/AsyncLoading.cpp
@@ -1498,7 +1498,7 @@ void FAsyncPackage::UpdateLoadPercentage()
 	}
 	else if (DeferredPostLoadObjects.Num() > 0)
 	{
-		NewLoadPercentage = DeferredPostLoadIndex / DeferredPostLoadObjects.Num();
+		NewLoadPercentage = static_cast<float>(DeferredPostLoadIndex) / DeferredPostLoadObjects.Num();
 	}
 	// It's also possible that we got so many objects to PostLoad that LoadPercantage will actually drop
 	LoadPercentage = FMath::Max(NewLoadPercentage, LoadPercentage);
